FEAT: ElementPropertyMetaData can now hold meta-data for properties of sub-items.
FEAT: `ElementPropertyMetaData` can now hold meta-data for properties of sub-items.
FEAT: Update `CloudRequestEngine` to be able to deserialize property meta-data with sub-properties.
TEST: Added a test to verify `CloudRequestEngine` behavior in this scenario.
FEAT: Added an interface `CloudRequestEngine,` and renamed implementing class to `CloudRequestEngineDefault`.
FEAT: `Pipeline.getElement` will now try to get any elements that can be cast to the supplied type if there are no direct matches. This allows it to get an element that implements a given interface.
TEST: Added tests to verify functionality in `CloudAspectEngineBase`